Adding verbose comment for remove when no files are present. Adding some doc.

+++ internal/cmd/remove.go
@@ -11,7 +11,14 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func removeFiles(_ *cobra.Command, _ []string) error {
+func removeFiles(cmd *cobra.Command, args []string) error {
+	if len(files) == 0 {
+		if verbose {
+			fmt.Printf("No source files were specified in a remove operation.")
+		}
+		return nil
+	}
+
 	for _, f := range files {
 		err := os.RemoveAll(f)
 		if err != nil {
